add randbats command and reorder gen autocomplete
